==> 01_hv_data_import.Rmd <==
---
title: 'HV Analysis - Data Import and Cleansing'
author: 'W241 Dan, Anu, Keane' 
output:
  rmarkdown::github_document
---

```{r,  echo=FALSE}
# load packages 
library(foreign)
library(data.table)
library(sandwich)
library(stargazer)
library(ggplot2)
library(gridExtra)

set.seed(42)
```

## Data Import and Cleansing

```{r import, echo=FALSE}
# Import participant data
d <- fread('../data/raw/participants.csv', header = TRUE, skip = 2) # Participant pre-treatment survey
answers <- fread("../data/raw/answers.csv", header = TRUE) # Assessment answers
a1 <- fread('../data/raw/assessment1.csv', header = TRUE, skip = 2) # Assessment 1 results
a2 <- fread('../data/raw/assessment2.csv', header = TRUE, skip = 2) # Assessment 2 results

# Compliance data
# group3 participated on days 1, 2, 3 and 7
g3_d1 <- fread('../data/raw/group3_day1.csv', header = TRUE, skip = 2)
g3_d2 <- fread('../data/raw/group3_day2.csv', header = TRUE, skip = 2)
g3_d3 <- fread('../data/raw/group3_day3.csv', header = TRUE, skip = 2)
g3_d7 <- fread('../data/raw/group3_day7.csv', header = TRUE, skip = 2)

# group4 participated on days 3, 5,6 and 7
g4_d3 <- fread('../data/raw/group4_day3.csv', header = TRUE, skip = 2)
g4_d5 <- fread('../data/raw/group4_day5.csv', header = TRUE, skip = 2)
g4_d6 <- fread('../data/raw/group4_day6.csv', header = TRUE, skip = 2)
g4_d7 <- fread('../data/raw/group4_day7.csv', header = TRUE, skip = 2)

```

### Data Cleansing: Pre-treatment Survey

```{r survey_cl, echo=FALSE}
# Fix headers
d_cols <- c('StartDate',	'EndDate',	'Status',	'IPAddress',	'Progress',	'Duration (in seconds)',	'Finished',	'RecordedDate',	'ResponseId',	'RecipientLastName',	'RecipientFirstName',	'RecipientEmail',	'ExternalReference',	'LocationLatitude',	'LocationLongitude',	'DistributionChannel',	'UserLanguage',	'browser',	'browser_version',	'os',	'resolution',	'name',	'email',	'age',	'education',	'num_languages',	'languages',	'got_familiar',	'hv_proficiency',	'Q9 - Topics')
setnames(d, old = names(d),new=d_cols)
d[, "email" := tolower(email)]

# Exclude incomplete entries
d <- d[Finished == TRUE, ]

# Remove unused columns
d <- d[ , -c("IPAddress","ResponseId","Duration (in seconds)","resolution","UserLanguage","DistributionChannel", "StartDate","EndDate","Status","RecordedDate","Progress","Finished","RecipientLastName", "RecipientFirstName", "RecipientEmail", "ExternalReference","Q9 - Topics")]

# Create pre-knowledge block
d[ , pre_hv := 0]
d[ hv_proficiency == "Intermediate (I can recognize a few words and phrases)", pre_hv := 1]
```

### Data Cleansing: Assessment 1 and 2

```{r assessments_cl, echo=FALSE}
# Fix headers
a_cols <- c('StartDate',	'EndDate',	'Status',	'IPAddress',	'Progress',	'Duration (in seconds)',	'Finished',	'RecordedDate',	'ResponseId',	'RecipientLastName',	'RecipientFirstName',	'RecipientEmail',	'ExternalReference',	'LocationLatitude',	'LocationLongitude',	'DistributionChannel',	'UserLanguage',	'browser',	'browser_version',	'os',	'resolution',	'name',	'email',	'q1_start',	'q1_lastclick',	'q1_submit',	'q1_clickct',	'q1_answer',	'q2_start',	'q2_lastclick',	'q2_submit',	'q2_clickct',	'q2_answer',	'q3_start',	'q3_lastclick',	'q3_submit',	'q3_clickct',	'q3_answer',	'q4_start',	'q4_lastclick',	'q4_submit',	'q4_clickct',	'q4_answer',	'q5_start',	'q5_lastclick',	'q5_submit',	'q5_clickct',	'q5_answer',	'q6_start',	'q6_lastclick',	'q6_submit',	'q6_clickct',	'q6_answer',	'q7_start',	'q7_lastclick',	'q7_submit',	'q7_clickct',	'q7_answer',	'q8_start',	'q8_lastclick',	'q8_submit',	'q8_clickct',	'q8_answer',	'q9_start',	'q9_lastclick',	'q9_submit',	'q9_clickct',	'q9_answer',	'q10_start',	'q10_lastclick',	'q10_submit',	'q10_clickct',	'q10_answer',	'q11_start',	'q11_lastclick',	'q11_submit',	'q11_clickct',	'q11_answer',	'q12_start',	'q12_lastclick',	'q12_submit',	'q12_clickct',	'q12_answer',	'q13_start',	'q13_lastclick',	'q13_submit',	'q13_clickct',	'q13_answer',	'q14_start',	'q14_lastclick',	'q14_submit',	'q14_clickct',	'q14_answer',	'q15_start',	'q15_lastclick',	'q15_submit',	'q15_clickct',	'q15_answer')
setnames(a1, old = names(a1),new=a_cols)

a1 <- a1[ , -c("IPAddress","ResponseId","Duration (in seconds)","resolution","UserLanguage","DistributionChannel", "StartDate","EndDate","Status","RecordedDate","Progress","Finished","RecipientLastName", "RecipientFirstName", "RecipientEmail", "ExternalReference")]

# Fix Headers
setnames(a2, old = names(a2),new=a_cols)
a2 <- a2[ , -c("IPAddress","ResponseId","Duration (in seconds)","resolution","UserLanguage","DistributionChannel", "StartDate","EndDate","Status","RecordedDate","Progress","Finished","RecipientLastName", "RecipientFirstName", "RecipientEmail", "ExternalReference")]

```


### Data Cleansing: Compliance - days 1, 2, 3, 5, 6, 7

```{r}
# Fix headers
c_cols <- c('startDate', 'endDate', 'status', 'ipAddress', 'progress', 'duration', 'finished', 'recordedDate', 'recordId', 'recipientLastName', 'recipientFirstName', 'recipientEmail', 'externalDataReference', 'locationLatitude', 'locationLongitude', 'distributionChannel', 'userLanguage', 'QID4_BROWSER', 'QID4_VERSION', 'QID4_OS', 'QID4_RESOLUTION', 'QID50_FIRST_CLICK', 'QID50_LAST_CLICK', 'QID50_PAGE_SUBMIT', 'QID50_CLICK_COUNT', 'QID51_TEXT')

#group3
setnames(g3_d1, old = names(g3_d1), new = c_cols)
setnames(g3_d2, old = names(g3_d2), new = c_cols)
setnames(g3_d3, old = names(g3_d3), new = c_cols)
setnames(g3_d7, old = names(g3_d7), new = c_cols)
# group4
setnames(g4_d3, old = names(g4_d3), new = c_cols)
setnames(g4_d5, old = names(g4_d5), new = c_cols)
setnames(g4_d6, old = names(g4_d6), new = c_cols)
setnames(g4_d7, old = names(g4_d7), new = c_cols)

```


### Export cleansed data to interim folder
```{r export, echo=FALSE}
fwrite(d, "../data/interim/participants_cl.csv")
fwrite(a1, "../data/interim/assessment1_cl.csv")
fwrite(a2, "../data/interim/assessment2_cl.csv")

fwrite(g3_d1, "../data/interim/group3_day1.csv")
fwrite(g3_d2, "../data/interim/group3_day2.csv")
fwrite(g3_d3, "../data/interim/group3_day3.csv")
fwrite(g3_d7, "../data/interim/group3_day7.csv")
fwrite(g4_d3, "../data/interim/group4_day3.csv")
fwrite(g4_d5, "../data/interim/group4_day5.csv")
fwrite(g4_d6, "../data/interim/group4_day6.csv")
fwrite(g4_d7, "../data/interim/group4_day7.csv")
```
