==> 02_hv_blocked_randomization.Rmd <==
---
title: 'HV Analysis - Blocked Randomization'
author: 'W241 Dan, Anu, Keane' 
output:
  rmarkdown::github_document
---

```{r,  echo=FALSE}
# load packages 
library(foreign)
library(data.table)
library(sandwich)
library(stargazer)
library(ggplot2)
library(gridExtra)

set.seed(42)
```

## Data Import

```{r import, echo=FALSE}
# Import participant data
d <- fread('../data/interim/participants_cl.csv', header = TRUE) # Participant pre-treatment survey
head(d)
```

## Blocked Randomization

```{r randomize, echo=FALSE}
d <- d[order(-pre_hv)]
pre_hv_obs <- sum(d[ ,pre_hv])
total_obs <- d[, .N]

randomize_blocked <- function(){
  c(sample(c(rep(0, round(pre_hv_obs/2)), rep(1, pre_hv_obs - round(pre_hv_obs/2)))), #group A
    sample(c(rep(0, round((total_obs- pre_hv_obs)/2)), rep(1, round((total_obs- pre_hv_obs)/2))))) #group B
}

# Assignment treatment
d <- d[ , treatment := randomize_blocked()]

# Check random assignment
table(d[ ,pre_hv], d[ ,treatment])
```

### Export cleansed data to interim folder

```{r export, echo=FALSE}
fwrite(d, "../data/interim/participants_cl_blkrand.csv")
```
