==> 04_hv_analysis.Rmd <==
---
title: 'HV Analysis - Analysis and Modeling'
author: 'W241 Dan, Anu, Keane' 
output:
  rmarkdown::github_document
---

```{r,  echo=FALSE}
# load packages 
library(foreign)
library(data.table)
library(sandwich)
library(stargazer)
library(ggplot2)
library(gridExtra)
library(cobalt)
library(MatchIt)

set.seed(42)
```

## Data Import

```{r import, echo=FALSE}
# Import participant data
dt <- fread('../data/processed/analysis.csv', header = TRUE) # Participant pre-treatment survey
compliers <- fread('../data/processed/complete_dataset.csv', header = TRUE) # Compliers
compliers <- compliers[dt, .(email, treatment, day1, day2, day3, day5, day6, day7), on='email'] # Filter to 24 experimental
```


### Covariate Balance

```{r}
# read the complete_dataset for covariate balance check
dc <- fread('../data/processed/complete_dataset.csv', header = TRUE) # Compliers
dc <- data.table(dc[ , c('treatment', 'pre_hv', 'age', 'education', 'num_languages')])

# Nearest neighbor matching with MatchIt
m.out <- matchit(treatment ~ pre_hv + age + education + num_languages, 
    data = dc)

# Checking balance before and after matching:
bal.tab(m.out, m.threshold = 0.1, un = TRUE)
```

```{r}

mod1 <- lm(treatment ~ 1 , data = dc)

mod2 <- lm(treatment ~ pre_hv + age + education + num_languages, data = dc)

stargazer(mod1,
          mod2,
          type = 'text')

anova(mod1, mod2, test = 'F')
```

The p-value from the F-test is not statisitically significant which indicates that the two linear regression models are similar.  Therefore the randomization of treatment assignment is indeed random.

```{r compliers, echo=FALSE}
m2s_ct <- dt[treatment == 1, .(n = .N)]$n
s2m_ct <- dt[treatment == 0, .(n = .N)]$n

m2s_alpha <- (compliers[treatment == 1, .(compliers = sum(day3))]$compliers + compliers[treatment == 1, .(compliers = sum(day5 + day6 + day7))]$compliers) / (m2s_ct * 4)

s2m_alpha <- (compliers[treatment == 0, .(compliers = sum(day1 + day2 + day3))]$compliers + compliers[treatment == 0, .(compliers = sum(day7))]$compliers) / (s2m_ct * 4)
```

## Calculate ATEs for scores
```{r, echo=FALSE}
# Treatment = 1 is Mass to Spaced
# Treatment = 0 is Spaced to Mass
# Observed ATEs include complier adjustment
o1 <- dt[treatment == 1, .("o1" = mean(a1_score))]$o1 #/ m2s_alpha
o2 <- dt[treatment == 1, .("o2" = mean(a2_score))]$o2 #/ m2s_alpha
o3 <- dt[treatment == 0, .("o3" = mean(a1_score))]$o3 #/ s2m_alpha
o4 <- dt[treatment == 0, .("o4" = mean(a2_score))]$o4 #/ s2m_alpha

# All taus are calculated as spaced - mass
tau1 <- (o3 - o1)
tau2 <- (o2 - o1)
tau3 <- (o3 - o4)
tau4 <- (o2 - o4)

cat("ATE with for Outcome = Scores\n")
cat("o1:", o1, "o2:", o2, "o3:", o3, "o4:", o4,"\n")
cat("tau1:", tau1, "tau2:", tau2, "tau3:", tau3, "tau4:", tau4)
```

## Calculate CACEs for scores
```{r, echo=FALSE}
# Treatment = 1 is Mass to Spaced
# Treatment = 0 is Spaced to Mass
# Observed ATEs include complier adjustment
o1 <- dt[treatment == 1, .("o1" = mean(a1_score))]$o1 / m2s_alpha
o2 <- dt[treatment == 1, .("o2" = mean(a2_score))]$o2 / m2s_alpha
o3 <- dt[treatment == 0, .("o3" = mean(a1_score))]$o3 / s2m_alpha
o4 <- dt[treatment == 0, .("o4" = mean(a2_score))]$o4 / s2m_alpha

# All taus are calculated as spaced - mass
tau1 <- (o3 - o1)
tau2 <- (o2 - o1)
tau3 <- (o3 - o4)
tau4 <- (o2 - o4)

cat("CACE with for Outcome = Scores\n")
cat("o1:", o1, "o2:", o2, "o3:", o3, "o4:", o4,"\n")
cat("tau1:", tau1, "tau2:", tau2, "tau3:", tau3, "tau4:", tau4)
```

# Calculate ATEs for times
```{r, echo=FALSE}
# Treatment = 1 is Mass to Spaced
# Treatment = 0 is Spaced to Mass
o1_t <- dt[treatment == 1, .("o1" = mean(a1_time))]$o1 #/ m2s_alpha
o2_t <- dt[treatment == 1, .("o2" = mean(a2_time))]$o2 #/ m2s_alpha
o3_t <- dt[treatment == 0, .("o3" = mean(a1_time))]$o3 #/ s2m_alpha
o4_t <- dt[treatment == 0, .("o4" = mean(a2_time))]$o4 #/ s2m_alpha

# All taus are calculated as spaced - mass
tau1_t <- o3_t - o1_t
tau2_t <- o2_t - o1_t
tau3_t <- o3_t - o4_t
tau4_t <- o2_t - o4_t

cat("ATE with for Outcome = Times\n")
cat("o1:", o1_t, "o2:", o2_t, "o3:", o3_t, "o4:", o4_t,"\n")
cat("tau1:", tau1_t, "tau2:", tau2_t, "tau3:", tau3_t, "tau4:", tau4_t)
```

# Calculate CACEs for times
```{r, echo=FALSE}
# Treatment = 1 is Mass to Spaced
# Treatment = 0 is Spaced to Mass
o1_t <- dt[treatment == 1, .("o1" = mean(a1_time))]$o1 / m2s_alpha
o2_t <- dt[treatment == 1, .("o2" = mean(a2_time))]$o2 / m2s_alpha
o3_t <- dt[treatment == 0, .("o3" = mean(a1_time))]$o3 / s2m_alpha
o4_t <- dt[treatment == 0, .("o4" = mean(a2_time))]$o4 / s2m_alpha

# All taus are calculated as spaced - mass
tau1_t <- o3_t - o1_t
tau2_t <- o2_t - o1_t
tau3_t <- o3_t - o4_t
tau4_t <- o2_t - o4_t

cat("CACE with for Outcome = Times\n")
cat("o1:", o1_t, "o2:", o2_t, "o3:", o3_t, "o4:", o4_t,"\n")
cat("tau1:", tau1_t, "tau2:", tau2_t, "tau3:", tau3_t, "tau4:", tau4_t)
```

# Cleansing data for regression
```{r}
dt1 <- dt[treatment == 1, .(browser, os, name, email, age, education, num_languages, pre_hv, "mass" = 1, "score" = a1_score, "time" = a1_time, "r2" = 0)] 
dt2 <- dt[treatment == 1, .(browser, os, name, email, age, education, num_languages, pre_hv, "mass" = 0, "score" = a2_score, "time" = a2_time, "r2" = 1)] 
dt3 <- dt[treatment == 0, .(browser, os, name, email, age, education, num_languages, pre_hv, "mass" = 0, "score" = a1_score, "time" = a1_time, "r2" = 0)] 
dt4 <- dt[treatment == 0, .(browser, os, name, email, age, education, num_languages, pre_hv, "mass" = 1, "score" = a2_score, "time" = a2_time, "r2" = 1)] 

dt_reg <- rbind(dt1, dt2, dt3, dt4, idcol = TRUE)
dt_reg[, "mobile" := 0]
dt_reg[os ==  "Android 8.0.0" |
       os ==  "iPhone" | 
       os == "iPad" | 
       os == "Android 9", "mobile" := 1]
```

```{r}
mod1 <- lm(score ~ mass, dt_reg)
mod2 <- lm(score ~ mass + mobile + factor(num_languages) + factor(education) + factor(pre_hv) + factor(email), dt_reg)
mod3 <- lm(score ~ mass + r2 + mobile + factor(num_languages) + factor(education) + factor(pre_hv) + factor(email), dt_reg)
stargazer(
  mod1,
  mod2,
  mod3,
  type = 'text',
  omit=c('num_languages', 'education', 'pre_hv', 'email'),
  header=F
  )

```

```{r}
mod1 <- lm(time ~ mass, dt_reg)
mod2 <- lm(time ~ mass + mobile + factor(num_languages) + factor(education) + factor(pre_hv) + factor(email), dt_reg)
mod3 <- lm(time ~ mass + r2 + mobile + factor(num_languages) + factor(education) + factor(pre_hv) + factor(email), dt_reg)
stargazer(
  mod1,
  mod2,
  mod3,
  type = 'text',
  omit=c('num_languages', 'education', 'pre_hv', 'email'),
  header=F
  )

```

## Create visualizations 

```{r, echo=FALSE}
# Assessment 1
labels <- c("Mass", "Spaced") 
names(labels) <- c(1, 0)

ms_score_hist1 <- ggplot(dt) +
  geom_histogram(aes(x=a1_score, fill=treatment), bins = 15, binwidth = 1)+
  facet_wrap(~treatment,labeller=labeller(treatment = labels), nrow=2) + xlab("Assessment 1 Score") + theme(legend.position = "none")

ms_score_density1 <- ggplot(dt) +
  geom_density(aes(x=a1_score, fill=treatment)) + 
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2)  + xlab("Assessment 1 Score") + theme(legend.position = "none")

sm_time_hist1 <- ggplot(dt) +
  geom_histogram(aes(x=a1_time, fill=treatment), bins = 15, binwidth = 1)+
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2) + xlab("Avg. Completion Time (Seconds)") + theme(legend.position = "none")

sm_time_density1 <- ggplot(dt) +
  geom_density(aes(x=a1_time, fill=treatment)) + 
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2) + xlab("Avg. Completion Time (Seconds)") + theme(legend.position = "none")

grid.arrange(ms_score_hist1, ms_score_density1, sm_time_hist1, sm_time_density1, nrow = 2)

# Assessment 2
labels <- c("Spaced", "Mass") 
names(labels) <- c(1, 0)

ms_score_hist2 <- ggplot(dt) +
  geom_histogram(aes(x=a2_score, fill=treatment), bins = 15, binwidth = 1)+
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2) + xlab("Assessment 2 Score") + theme(legend.position = "none")

ms_score_density2 <- ggplot(dt) +
  geom_density(aes(x=a2_score, fill=treatment)) + 
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2)  + xlab("Assessment 2 Score") + theme(legend.position = "none")

sm_time_hist2 <- ggplot(dt) +
  geom_histogram(aes(x=a2_time, fill=treatment), bins = 15, binwidth = 1)+
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2) + xlab("Avg. Completion Time (Seconds)") + theme(legend.position = "none")

sm_time_density2 <- ggplot(dt) +
  geom_density(aes(x=a2_time, fill=treatment)) + 
  facet_wrap(~treatment,labeller=labeller(treatment = labels),nrow=2)  + xlab("Avg. Completion Time (Seconds)") + theme(legend.position = "none")

grid.arrange(ms_score_hist2, ms_score_density2, sm_time_hist2, sm_time_density2, nrow = 2)
```

### Other visualizations
```{r,echo=FALSE}
browser <- ggplot(dt[, .(n = .N, browser)], aes(x="", y=n, fill=browser)) + geom_bar(width = 1, stat = "identity") + coord_polar("y", start=0)
age <- ggplot(dt[, .(n = .N, age)], aes(x="", y=n, fill=age)) + geom_bar(width = 1, stat = "identity") + coord_polar("y", start=0)
education <- ggplot(dt[, .(n = .N, education)], aes(x="", y=n, fill=education)) + geom_bar(width = 1, stat = "identity") + coord_polar("y", start=0)
num_languages <- ggplot(dt[, .(n = .N, num_languages)], aes(x="", y=n, fill=num_languages)) + geom_bar(width = 1, stat = "identity") + coord_polar("y", start=0)
got_familiar <- ggplot(dt[, .(n = .N, got_familiar)], aes(x="", y=n, fill=got_familiar)) + geom_bar(width = 1, stat = "identity") + coord_polar("y", start=0)
hv_proficiency <- ggplot(dt[, .(n = .N, hv_proficiency)], aes(x="", y=n, fill=hv_proficiency)) + geom_bar(width = 1, stat = "identity") + coord_polar("y", start=0)

browser
age
education
num_languages
got_familiar
hv_proficiency
#grid.arrange(browser, age, education, num_languages,got_familiar, hv_proficiency, nrow = 6)
#create block for mobile (1 if mobile)
```
