==> 05_hv_prizes.Rmd <==
---
title: 'HV Prizes - Raffle for Amazon gift cards'
author: 'W241 Dan, Anu, Keane' 
output:
  rmarkdown::github_document
---

```{r,  echo=FALSE}
# load packages 
library(data.table)

set.seed(42)
```

## Data Import

```{r import, echo=FALSE}
# Import participant data
d <- fread('../data/processed/analysis.csv', header = TRUE) # Participant pre-treatment survey
d <- data.table(d[ ,c('name', 'email')])
```

## Raffle

```{r}
# Select ten persons for $25 amazon gift card - random assignment of 1.  The remaining fourteen will get a $10 amazon gift card
d$prize1 <- sample(c(rep(1, 10), rep(0, 14)))
#d$prize2 <- rnorm(24, mean = 0, sd = 1)
```
